:art: Reorganizes project classes in different files.
